+++ UID_message.h
@@ -35,14 +35,14 @@ typedef struct {
 
 // user side functions
 int UID_createChannel(char *destMachine, UID_ClientChannelCtx *ctx);
-int UID_formatReqMsg(char *sender, int method, char *params, uint8_t *msg, size_t *size, int *sID);
-int UID_parseRespMsg(uint8_t *msg, size_t size, char *sender, size_t ssize, int *error, char *result, size_t rsize, int *sID);
+int UID_formatReqMsg(char *sender, int method, char *params, uint8_t *msg, size_t *size, int64_t *sID);
+int UID_parseRespMsg(uint8_t *msg, size_t size, char *sender, size_t ssize, int *error, char *result, size_t rsize, int64_t *sID);
 int UID_closeChannel(UID_ClientChannelCtx *ctx);
 
 // provider side functions
 int UID_accept_channel(uint8_t *in_msg, size_t in_size, UID_ServerChannelCtx *channel_ctx, uint8_t *first_msg, size_t *out_size);
-int UID_parseReqMsg(uint8_t *msg, size_t size, char *sender, size_t ssize, int *method, char *params, size_t psize, int *sID);
-int UID_formatRespMsg(char *sender, char *result, int error, int sID, uint8_t *msg, size_t *size);
+int UID_parseReqMsg(uint8_t *msg, size_t size, char *sender, size_t ssize, int *method, char *params, size_t psize, int64_t *sID);
+int UID_formatRespMsg(char *sender, char *result, int error, int64_t sID, uint8_t *msg, size_t *size);
 int UID_closeServerChannel(UID_ServerChannelCtx *ctx);
